requirements.txt
Flask
flask-cors
google-generativeai
langchain
langchain-google-genai
python-dotenv
langchain-community
langchain-text-splitters
gunicorn